+++ HISTORY.rst
@@ -4,6 +4,7 @@ Release history
 Current release
 ---------------
 
+* Deprecation warning: support for py2.6 and py3.3 will be dropped
 * Bugfixes and improvements
 * Localisation updates